requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
